Add utils

+++ utils/install-helm.sh
@@ -6,11 +6,16 @@ cd "$this_dir" || exit 1
 . ./env.sh
 
 function install_helm() {
-  log_info "Installing Helm"
-  curl -fsSL -o get_helm.sh https://raw.githubusercontent.com/helm/helm/main/scripts/get-helm-3
-  chmod 700 get_helm.sh
-  ./get_helm.sh
-  rm get_helm.sh
+  if [ "$(uname)" == "Darwin" ]; then
+    log_info "Installing Helm using Brew."
+    brew install helm
+  else
+    log_info "Installing Helm."
+    curl -fsSL -o get_helm.sh https://raw.githubusercontent.com/helm/helm/main/scripts/get-helm-3
+    chmod 700 get_helm.sh
+    ./get_helm.sh
+    rm get_helm.sh
+  fi
 }
 
 install_helm
